overwrite with original 2.0.3 src files

+++ src/playout.h
@@ -126,7 +126,7 @@ double Score(Board& b, double komi=KOMI);
 int Playout(Board& b, double komi=KOMI);
 int PlayoutRandom(Board&b, double komi=KOMI);
 
-double PlayoutLGR(Board& b, LGR& lgr, double komi=KOMI);
-double PlayoutLGR(Board& b, LGR& lgr, Statistics& stat, double komi=KOMI);
+int PlayoutLGR(Board& b, LGR& lgr, double komi=KOMI);
+int PlayoutLGR(Board& b, LGR& lgr, Statistics& stat, double komi=KOMI);
 
 extern bool japanese_rule;
